fix: preserve MediaType param case + bound Digest nonce counter (H3, H4)
MediaType.parse() was lowercasing every parameter value, including
boundary= and base64 payloads. Multipart bodies with any uppercase
letter in the boundary were silently corrupted on parse/round-trip
(H3). Values are now case-preserved; the charset getter still matches
case-insensitively where the spec requires it.

DigestChallengeHandler used AtomicInteger for the nonce counter,
producing a negative %08x at 2^31 requests (H4). Bumped to AtomicLong;
no realistic lifetime hits the new ceiling.

Other items addressed in the same HTTP-model / I/O files:
- M7: MediaType.parse() now unquotes and unescapes quoted parameter
  values (charset="utf-8", boundary="abc\\\"def").
- M2: Status.fromCodeOrNull / fromCode use a class-load LOOKUP map;
  the previous O(n) entries scan sat on the hot path.
- M9: Io.withProvider moves out of the public Io object into a new
  sdk-core testFixtures module (org.dexpace.sdk.core.testing.withProvider).
  The seam was always test-only and unsafe for parallel test classes;
  sdk-io-okio3 picks it up via testFixtures(project(":sdk-core")).
- M20: drops the redundant Headers.builder(headers) static; callers use
  headers.newBuilder().
- M21: BasicChallengeHandler accepts empty username or password
  (RFC 7617 permits "user:" and ":password") via isNotEmpty.
- M8: PagedIterable KDoc documents that the Long.MAX_VALUE default is
  unbounded-by-design; production callers should cap it explicitly.

+++ sdk-core/src/main/kotlin/org/dexpace/sdk/core/http/auth/DigestChallengeHandler.kt
@@ -7,7 +7,7 @@ import java.nio.charset.Charset
 import java.security.MessageDigest
 import java.security.NoSuchAlgorithmException
 import java.util.Locale
-import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.atomic.AtomicLong
 
 /**
  * RFC 7616 `Digest` authentication handler.
@@ -24,7 +24,7 @@ import java.util.concurrent.atomic.AtomicInteger
  * legacy MD5.
  *
  * Thread-safety: safe across concurrent requests. The nonce counter is an
- * [AtomicInteger]; `MessageDigest` is not thread-safe and is therefore instantiated
+ * [AtomicLong]; `MessageDigest` is not thread-safe and is therefore instantiated
  * per call (microsecond cost, acceptable per spec).
  *
  * If the JVM lacks `MD5` or `SHA-256` (should never happen), the handler throws
@@ -42,7 +42,7 @@ class DigestChallengeHandler @JvmOverloads constructor(
         require(preferredAlgorithms.isNotEmpty()) { "preferredAlgorithms must not be empty" }
     }
 
-    private val nonceCount = AtomicInteger(0)
+    private val nonceCount = AtomicLong(0)
 
     override fun canHandle(challenges: List<AuthenticateChallenge>): Boolean =
         pickChallenge(challenges) != null
@@ -231,7 +231,7 @@ class DigestChallengeHandler @JvmOverloads constructor(
     }
 
     /** Pads [count] to exactly 8 lower-case hex digits per RFC 7616 §3.4. */
-    private fun ncHex(count: Int): String = String.format(Locale.US, "%08x", count)
+    private fun ncHex(count: Long): String = String.format(Locale.US, "%08x", count)
 
     /** Non-blocking cnonce: 16 hex chars (high 64 bits of a fresh type-4 UUID). */
     private fun generateCnonce(): String {
